backend/requirements.txt
gradio
transformers  # For Hugging Face models
torch
torchvision
pillow        # For image processing
python-dotenv # For loading environment variables
timm
